doMC-package {doMC}	R Documentation

The doMC Package

Description

The doMC package provides a parallel backend for the foreach/%dopar% function using the multicore functionality of the parallel package.

Details

Further information is available in the following help topics:

registerDoMC 	 register doMC to be used by foreach/%dopar%

To see a tutorial introduction to the doMC package, use vignette("gettingstartedMC"). To see a tutorial introduction to the foreach package, use vignette("foreach").

To see a demo of doMC computing the sinc function, use demo(sincMC).

Some examples (in addition to those in the help pages) are included in the “examples” directory of the doMC package. To list the files in the examples directory, use list.files(system.file("examples", package="doMC")). To run the bootstrap example, use source(system.file("examples", "bootMC.R", package="doMC")). This is a simple benchmark, executing both sequentally and in parallel. There are many more examples that come with the foreach package, which will work with the doMC package if it is registered as the parallel backend.

For a complete list of functions with individual help pages, use library(help="doMC").

[Package doMC version 1.3.3 Index]