registerDoMC {doMC}	R Documentation

registerDoMC

Description

The registerDoMC function is used to register the multicore parallel backend with the foreach package.

Usage

registerDoMC(cores=NULL, ...)

Arguments

cores
The number of cores to use for parallel execution. If not specified, the number of cores is set to the value of options("cores"), if specified, or to approximately half the number of cores detected by the parallel package.

...
Package options. Currently, only the nocompile option is supported. If nocompile is set to TRUE, compiler support is disabled.

Details

The multicore functionality, originally written by Simon Urbanek and subsumed in the parallel package in R 2.14.0, provides functions for parallel execution of R code on machines with multiple cores or processors, using the system fork call to spawn copies of the current process. The multicore functionality, and therefore registerDoMC, should not be used in a GUI environment, because multiple processes then share the same GUI.

[Package doMC version 1.3.3 Index]